feat: :art: bun add react label and moving files for a better structure
