+++ package.json
@@ -31,5 +31,8 @@
 		"core-js-pure",
 		"esbuild",
 		"sharp"
-	]
+	],
+	"dependencies": {
+		"@radix-ui/react-label": "^2.1.1"
+	}
 }
